refactoring of fs format : - removed need for error-prone format option of type Object in FileSystemType.format - extracted format command from FileSystemType (which has no sense for lot of fs) -> created interface Formatter instead - splitted format commands per filesystem (simplify the syntax of these commands) - created an abstract format command that do common parts of formatting job
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@3668 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ fs/src/fs/org/jnode/fs/ramfs/RAMFileSystem.java
@@ -9,35 +9,35 @@ import org.jnode.fs.FileSystemException;
 
 /**
  * A Filesystem implementation in the system RAM.
- * 
+ *
  * @author peda
  */
 public class RAMFileSystem implements FileSystem {
 
 	private Device device;
 	private boolean readOnly;
-	
+
 	private RAMDirectory root;
 
 	private final long maxSize;
 	private long summedBufferSize;
 	private long summedFileSize;
-	
+
 	/**
 	 * Constructor for RAMFileSystem
-	 * @param device 
-	 * @param readOnly 
-	 * @param maxSize 
-	 * @throws FileSystemException 
+	 * @param device
+	 * @param readOnly
+	 * @param maxSize
+	 * @throws FileSystemException
 	 */
-	public RAMFileSystem(Device device, boolean readOnly, int maxSize) throws FileSystemException {
+	public RAMFileSystem(Device device, boolean readOnly, long maxSize) throws FileSystemException {
 		this.device = device;
 		this.readOnly = readOnly;
-		
+
 		this.maxSize = maxSize;
 		summedBufferSize = 0;
 		summedFileSize = 0;
-		
+
 		root = new RAMDirectory(this, null, "");
 	}
 
@@ -58,7 +58,7 @@ public class RAMFileSystem implements FileSystem {
 			throw new IOException("Filesystem closed");
 		return root;
 	}
-	
+
 	/**
 	 * (non-Javadoc)
 	 * @see org.jnode.fs.FileSystem#isReadOnly()
@@ -90,7 +90,7 @@ public class RAMFileSystem implements FileSystem {
     public long getTotalSpace() {
     	return maxSize;
     }
-    
+
     /**
      * (non-Javadoc)
      * @see org.jnode.fs.FileSystem#getFreeSpace()
@@ -98,7 +98,7 @@ public class RAMFileSystem implements FileSystem {
     public long getFreeSpace() {
     	return maxSize - summedBufferSize;
     }
-    
+
     /**
      * (non-Javadoc)
      * @see org.jnode.fs.FileSystem#getUsableSpace()
@@ -106,7 +106,7 @@ public class RAMFileSystem implements FileSystem {
     public long getUsableSpace() {
     	return maxSize - summedFileSize;
     }
-    
+
 	synchronized void addSummmedBufferSize(long toAdd) {
 		summedBufferSize += toAdd;
 	}
